Move lead school sync logic to a concern
This means that we can add this logic for any type in the app that needs migrating from lead_school to lead_partner regardless of any variation in the column name.
