+++ spec/models/bulk_update/recommendations_upload_row_spec.rb
@@ -10,10 +10,6 @@ RSpec.describe BulkUpdate::RecommendationsUploadRow do
   end
 
   describe "syncing lead_partner and lead_school" do
-    before do
-      BulkUpdate::RecommendationsUploadRow.set_lead_columns(:lead_school, :lead_partner)
-    end
-
     let(:row) { create(:bulk_update_recommendations_upload_row) }
 
     context "when lead_partner is changed" do
